change from npm utility to meteor package
easier than using ant to keep track of which commands do what

hack that uses environment variables to add commands to the meteor tool

converted most tests to Tinytest
(great test framework)

called the tinytest callback argument 'assert' instead of 'test' so that I
could follow changes easier in git diff.

exported an object with all the different methods for testing.  I guess it
could come in handy at some point.  I probably could have left the old
names for the objects, and just included the files in the tests instead
of the module and not exported anything.

Now it extracts the sizes from the .meteor library at run time instead of
just testing the hard coded values.  That MAY allow it to stay up to date
for longer.

also removed tests for execSync - they were just there to figure out how
it worked.

globally set 'require' to Npm.require so that we don't have to change all
the npm imports right away